Move function to http.c
